+++ src/ui/eventlistview.h
@@ -48,8 +48,10 @@ public:
 
 protected:
     bool eventFilter(QObject *watched, QEvent *event) override;
+    void changeEvent(QEvent *event) override;
 
 private:
+    void applyRowIndexFont();
     void rebuildChunkCombo();
     void chunkPicked(int comboIndex);
     void filterChanged();